Aggiunto timer per regolare il controllo dello status register ad una frequenza minore. Pulizia del codice. Da sistemare README.md

+++ ANDRIOTTO_LUCA.cydsn/accelerometer.c
@@ -18,13 +18,15 @@ ErrorCode I2C_LIS3DH_Get_Raw_Data(int16_t* data);
 
 ErrorCode I2C_LIS3DH_Start()
 {
-    // Reading EEPROM register dedicated to startup ODR
-    // Checking if the memorized value is in allowed range
-    // if the value is out of range --> ODR is set to 1Hz
+    // Reading EEPROM register dedicated to startup ODR.
+    // Checking if the memorized value is in allowed range.
+    // if the value is out of range --> ODR is set to the default initial value.
     if(EEPROM_ReadByte(EEPROM_REGISTER) > EEPROM_FINAL_VALUE)
     {
-        EEPROM_UpdateTemperature();
-        EEPROM_WriteByte(EEPROM_INIT_VALUE,EEPROM_REGISTER);
+        // Updating temperature before writing the initial value.
+		EEPROM_UpdateTemperature();
+		// Initialization of the EEPROM register
+        EEPROM_WriteByte(EEPROM_INIT_VALUE, EEPROM_REGISTER);
     }
     
     // init CTRL_REG1
@@ -47,26 +49,26 @@ ErrorCode I2C_LIS3DH_Manage_Data(outtype* array)
     uint8_t out[BYTE_TO_TRANSFER];
     out[0] = 0xA0;                      // --> HEADER
     out[BYTE_TO_TRANSFER - 1] = 0xC0;   // --> TAIL
-
+    
     // Array dedicated to save accelerometer data in digit.
     int16_t sensorData[3];
-        
     
     // Data acquisition
     I2C_LIS3DH_Get_Raw_Data(sensorData);
     
     // Data conversion
-    for(int i = 0; i < LIS3DH_OUT_AXES; i++)  
-	array[i].data = sensorData[i] * CONVERSION; // populating the array's float data.
+    for(uint8_t i = 0; i < LIS3DH_OUT_AXES; i++)  
+	    array[i].data = sensorData[i] * CONVERSION; // populating the array's float data.
     
     // Counters
-    int arrayIndex=0, shiftIndex=0;    // arrayIndex is used to browse the array cells. 
-                                       // shiftIndex is used in order to perform the right shift operation on the data.
-    for(int count = 1; count < BYTE_TO_TRANSFER - 1; count++) 
+    uint8_t arrayIndex=0,    	// arrayIndex is used to browse the array cells. 
+			shiftIndex=0;   	// shiftIndex is used in order to perform the right shift operation on the data.
+                                       
+    for(uint8_t count = 1; count < BYTE_TO_TRANSFER - 1; count++) 
     {
-        out[count] = array[arrayIndex].mask>>8*shiftIndex & 0xff;   // populating the buffer with binary translated float properly masked.
-        shiftIndex=(shiftIndex+1) % BYTES_PER_AXIS;                 // updating the shift index.
-        if(!(count % BYTES_PER_AXIS)) arrayIndex++;                 // updating the array index, once all the data related to one axis are gathered.
+        out[count] = array[arrayIndex].mask >> 8 * shiftIndex & 0xff;   // populating the buffer with binary translated float properly masked.
+        shiftIndex = (shiftIndex+1) % BYTES_PER_AXIS;                 	// updating the shift index.
+        if(!(count % BYTES_PER_AXIS)) arrayIndex++;                 	// updating the array index, once all the data related to one axis are gathered.
     }
     
     // Sending data buffer
@@ -77,13 +79,13 @@ ErrorCode I2C_LIS3DH_Manage_Data(outtype* array)
 
 ErrorCode I2C_LIS3DH_Get_Raw_Data(int16_t* data)
 {
-    uint8_t sensorData[LIS3DH_OUT_N]; // This array is storing left aligned bytes coming from accelerometer
+    uint8_t sensorData[LIS3DH_OUT_N]; // This array is storing left-aligned bytes coming from accelerometer
     
     // Storing data in temporary array.
     I2C_Peripheral_ReadRegisterMulti(LIS3DH_DEVICE_ADDRESS, LIS3DH_OUT_X_L, LIS3DH_OUT_N, sensorData);
     
-    // Re-arranging data to have them right aligned as integers
-    for(int i = 0; i < LIS3DH_OUT_AXES; i++) data[i] = (int16_t)((sensorData[2*i+1]<<8 | sensorData[2*i])) >> LIS3DH_RIGHT_SHIFT;   
+    // Re-arranging data to have them right-aligned as integers
+    for(int i = 0; i < LIS3DH_OUT_AXES; i++) data[i] = (int16_t)((sensorData[2*i+1] << 8 | sensorData[2*i])) >> LIS3DH_RIGHT_SHIFT;   
     
     return NO_ERROR;
 }
